more firecracker adjustments
