+++ pkg/virtualizers/firecracker/virtualizer_linux.go
@@ -8,12 +8,10 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"syscall"
@@ -247,119 +245,6 @@ type operation struct {
 	ctx    context.Context
 }
 
-// writeCounter counts the number of bytes written to it.
-type writeCounter struct {
-	total      int64 // total size
-	downloaded int64 // downloaded # of bytes transferred
-	onProgress func(downloaded int64, total int64)
-}
-
-// Write implements the io.Writer interface.
-//
-// Always completes and never returns an error.
-func (wc *writeCounter) Write(p []byte) (n int, e error) {
-	n = len(p)
-	wc.downloaded += int64(n)
-	wc.onProgress(wc.downloaded, wc.total)
-	return
-}
-func newWriter(size int64, onProgress func(downloaded, total int64)) io.Writer {
-	return &writeCounter{total: size, onProgress: onProgress}
-}
-
-// fetchVMLinux reads the kernel it wants to run and returns the vm linux required to run
-// Will download the vmlinux if it doesn't exist
-func (o *operation) fetchVMLinux(kernel string) (string, error) {
-	o.updateStatus(fmt.Sprintf("Fetching VMLinux searching %s for %s", o.firecrackerPath, kernel))
-	// check if vmlinux is on system at valid path
-	_, err := os.Stat(filepath.Join(o.firecrackerPath, kernel))
-	if err != nil {
-		// file doesn't exist must download from bucket
-		o.updateStatus(fmt.Sprintf("VMLinux for kernel doesn't exist downloading..."))
-		// Download vmlinux from google
-		url := DownloadPath + kernel
-		client := http.DefaultClient
-		// Create file locally to download
-		file, err := os.Create(filepath.Join(o.firecrackerPath, kernel))
-		if err != nil {
-			return "", err
-		}
-		defer file.Close()
-
-		// Determinate the file size
-		resp, err := client.Head(url)
-		if err != nil {
-			os.Remove(file.Name())
-			return "", err
-		}
-		if resp.StatusCode == 404 {
-			os.Remove(file.Name())
-			return "", fmt.Errorf("Kernel '%s' VMLinux does not exist", kernel)
-		}
-		contentLength := resp.Header.Get("content-length")
-		length, err := strconv.Atoi(contentLength)
-		if err != nil {
-			os.Remove(file.Name())
-			return "", err
-		}
-
-		// Make request
-		resp, err = client.Get(url)
-		if err != nil {
-			os.Remove(file.Name())
-			return "", err
-		}
-		defer resp.Body.Close()
-		p := o.logger.NewProgress("Downloading VMLinux", "Bytes", int64(length))
-		defer p.Finish(false)
-		// pipe stream
-		var pDownloaded = int64(0)
-		body := io.TeeReader(resp.Body, newWriter(int64(length), func(downloaded, total int64) {
-			p.Increment(downloaded - pDownloaded)
-			pDownloaded = downloaded
-		}))
-		_, err = io.Copy(file, body)
-		if err != nil {
-			os.Remove(file.Name())
-			return "", err
-		}
-
-	}
-
-	return filepath.Join(o.firecrackerPath, kernel), nil
-}
-
-// log writes a log to the channel for the job
-func (o *operation) log(text string, v ...interface{}) {
-	o.Logs <- fmt.Sprintf(text, v...)
-}
-
-// finished finishes the job and cleans up the channels
-func (o *operation) finished(err error) {
-	o.finishedLock.Lock()
-	defer o.finishedLock.Unlock()
-	if o.isFinished {
-		return
-	}
-
-	o.isFinished = true
-	if err != nil {
-		o.Logs <- fmt.Sprintf("Error: %v", err)
-		o.Status <- fmt.Sprintf("Failed: %v", err)
-		o.Error <- err
-	}
-
-	close(o.Logs)
-	close(o.Status)
-	close(o.Error)
-}
-
-// updateStatus updates the status of the job to provide more feedback to the user
-func (o *operation) updateStatus(text string) {
-	o.Status <- text
-	o.Logs <- text
-}
-
 // Serial returns the serial logger which contains the serial output of the application
 func (v *Virtualizer) Serial() *logger.Logger {
 	return v.serialLogger
@@ -454,45 +339,6 @@ func (v *Virtualizer) Close(force bool) error {
 	return nil
 }
 
-// ConvertToVM is a wrapper function that provides us abilities to use the old APIs
-func (v *Virtualizer) ConvertToVM() interface{} {
-	info := v.config.Info
-	vm := v.config.VM
-	system := v.config.System
-	programs := make([]virtualizers.ProgramSummaries, 0)
-
-	for _, p := range v.config.Programs {
-		programs = append(programs, virtualizers.ProgramSummaries{
-			Binary: p.Binary,
-			Args:   string(p.Args),
-			Env:    p.Env,
-		})
-	}
-
-	machine := &virtualizers.VirtualMachine{
-		ID:       v.name,
-		Author:   info.Author,
-		CPUs:     int(vm.CPUs),
-		RAM:      vm.RAM,
-		Disk:     vm.DiskSize,
-		Created:  v.created,
-		Date:     info.Date.Time(),
-		Networks: v.routes,
-		Kernel:   vm.Kernel,
-		Name:     info.Name,
-		Summary:  info.Summary,
-		Source:   v.source.(virtualizers.Source),
-		URL:      string(info.URL),
-		Version:  info.Version,
-		Programs: programs,
-		Hostname: system.Hostname,
-		Platform: v.pname,
-		Status:   v.state,
-	}
-
-	return machine
-}
-
 // Prepare prepares the virtualizer with the appropriate fields to run the virtualizer
 func (v *Virtualizer) Prepare(args *virtualizers.PrepareArgs) *virtualizers.VirtualizeOperation {
 
@@ -537,39 +383,82 @@ type firecrackerFormatter struct {
 	logrus.TextFormatter
 }
 
+// Format writes the message differently compared to logsrus
 func (f *firecrackerFormatter) Format(entry *logrus.Entry) ([]byte, error) {
 	return []byte(entry.Message), nil
 }
 
-// prepare sets the fields and arguments to spawn the virtual machine
-func (o *operation) prepare(args *virtualizers.PrepareArgs) {
-	var returnErr error
-
+func (o *operation) initializeVM(args *virtualizers.PrepareArgs) error {
 	o.updateStatus(fmt.Sprintf("Building firecracker command and tap interfaces..."))
-	defer func() {
-		o.finished(returnErr)
-	}()
 
 	o.state = "initializing"
 	o.name = args.Name
 	err := os.MkdirAll(args.FCPath, os.ModePerm)
 	if err != nil {
-		returnErr = err
-		return
+		return err
 	}
 	o.folder = filepath.Dir(args.ImagePath)
 	o.id = strings.Split(filepath.Base(o.folder), "-")[1]
+
+	o.gctx = context.Background()
+	o.vmmCtx, o.vmmCancel = context.WithCancel(o.gctx)
+
+	o.kip, err = o.fetchVMLinux(fmt.Sprintf("firecracker-%s", o.config.VM.Kernel))
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// prepare sets the fields and arguments to spawn the virtual machine
+func (o *operation) prepare(args *virtualizers.PrepareArgs) {
+	var returnErr error
+	defer func() {
+		o.finished(returnErr)
+	}()
+
 	diskpath := filepath.ToSlash(args.ImagePath)
 
+	err := o.initializeVM(args)
+	if err != nil {
+		returnErr = err
+		return
+	}
+	err = o.sendDeviceCreation()
+	if err != nil {
+		returnErr = err
+		return
+	}
+
+	fcCfg, machineOpts := o.generateFirecrackerConfig(diskpath)
+	// append new fields to overarching struct
+	o.machineOpts = machineOpts
+	o.fconfig = fcCfg
+
+	o.state = "ready"
+
+	_, loaded := virtualizers.ActiveVMs.LoadOrStore(o.name, o.Virtualizer)
+	if loaded {
+		returnErr = errors.New("virtual machine already exists")
+		return
+	}
+
+	if args.Start {
+		err = o.Start()
+		if err != nil {
+			returnErr = err
+			return
+		}
+	}
+}
+
+func (o *operation) generateFirecrackerConfig(diskpath string) (firecracker.Config, []firecracker.Opt) {
 	logger := log.New()
 	logger.SetFormatter(&firecrackerFormatter{log.TextFormatter{
 		DisableColors: false,
 		ForceColors:   true,
 	}})
 	logger.Out = o
-
-	ctx := context.Background()
-	vmmCtx, vmmCancel := context.WithCancel(ctx)
 	devices := []models.Drive{}
 
 	rootDrive := models.Drive{
@@ -581,13 +470,35 @@ func (o *operation) prepare(args *virtualizers.PrepareArgs) {
 	}
 
 	devices = append(devices, rootDrive)
+	var interfaces []firecracker.NetworkInterface
 
-	o.kip, err = o.fetchVMLinux(fmt.Sprintf("firecracker-%s", o.config.VM.Kernel))
-	if err != nil {
-		returnErr = err
-		return
+	for i := 0; i < len(o.tapDevice.Devices); i++ {
+		interfaces = append(interfaces,
+			firecracker.NetworkInterface{
+				StaticConfiguration: &firecracker.StaticNetworkConfiguration{
+					HostDevName: o.tapDevice.Devices[i],
+				},
+			},
+		)
 	}
+	return firecracker.Config{
+			SocketPath:      filepath.Join(o.folder, fmt.Sprintf("%s.%s", o.name, "socket")),
+			KernelImagePath: o.kip,
+			KernelArgs:      fmt.Sprintf("init=/vorteil/vinitd root=PARTUUID=%s reboot=k panic=1 pci=off vt.color=0x00", vimg.Part2UUIDString),
+			Drives:          devices,
+			MachineCfg: models.MachineConfiguration{
+				VcpuCount:  firecracker.Int64(int64(o.config.VM.CPUs)),
+				HtEnabled:  firecracker.Bool(false),
+				MemSizeMib: firecracker.Int64(int64(o.config.VM.RAM.Units(vcfg.MiB))),
+			},
+			NetworkInterfaces: interfaces,
+			ForwardSignals:    []os.Signal{},
+		}, []firecracker.Opt{
+			firecracker.WithLogger(log.NewEntry(logger)),
+		}
+}
 
+func (o *operation) sendDeviceCreation() error {
 	cd := CreateDevices{
 		Id:     o.id,
 		Routes: len(o.routes),
@@ -595,79 +506,28 @@ func (o *operation) prepare(args *virtualizers.PrepareArgs) {
 
 	cdm, err := json.Marshal(cd)
 	if err != nil {
-		returnErr = err
-		return
+
+		return err
 	}
 	resp, err := http.Post("http://localhost:7476/", "application/json", bytes.NewBuffer(cdm))
 	if err != nil {
-		returnErr = errors.New("Run 'sudo vorteil firecracker-setup' for the listener")
-		return
+		return errors.New("Run 'sudo vorteil firecracker-setup' for the listener")
 	}
 
 	defer resp.Body.Close()
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		returnErr = err
-		return
+		return err
 	}
 	var ifs Devices
 	err = json.Unmarshal(body, &ifs)
 	if err != nil {
-		returnErr = err
-		return
+
+		return err
 	}
 
 	o.tapDevice = ifs
-	var interfaces []firecracker.NetworkInterface
-
-	for i := 0; i < len(ifs.Devices); i++ {
-		interfaces = append(interfaces,
-			firecracker.NetworkInterface{
-				StaticConfiguration: &firecracker.StaticNetworkConfiguration{
-					HostDevName: ifs.Devices[i],
-				},
-			},
-		)
-	}
-
-	fcCfg := firecracker.Config{
-		SocketPath:      filepath.Join(o.folder, fmt.Sprintf("%s.%s", o.name, "socket")),
-		KernelImagePath: o.kip,
-		KernelArgs:      fmt.Sprintf("loglevel=9 init=/vorteil/vinitd root=PARTUUID=%s reboot=k panic=1 pci=off vt.color=0x00", vimg.Part2UUIDString),
-		Drives:          devices,
-		MachineCfg: models.MachineConfiguration{
-			VcpuCount:  firecracker.Int64(int64(o.config.VM.CPUs)),
-			HtEnabled:  firecracker.Bool(false),
-			MemSizeMib: firecracker.Int64(int64(o.config.VM.RAM.Units(vcfg.MiB))),
-		},
-		NetworkInterfaces: interfaces,
-		ForwardSignals:    []os.Signal{},
-	}
-
-	machineOpts := []firecracker.Opt{
-		firecracker.WithLogger(log.NewEntry(logger)),
-	}
-
-	// append new fields to overarching struct
-	o.gctx = ctx
-	o.vmmCtx = vmmCtx
-	o.vmmCancel = vmmCancel
-	o.machineOpts = machineOpts
-	o.fconfig = fcCfg
-
-	o.state = "ready"
-
-	_, loaded := virtualizers.ActiveVMs.LoadOrStore(o.name, o.Virtualizer)
-	if loaded {
-		returnErr = errors.New("virtual machine already exists")
-	}
-
-	if args.Start {
-		err = o.Start()
-		if err != nil {
-			returnErr = err
-		}
-	}
+	return nil
 }
 
 // Start create the virtualmachine and runs it
